PRODUCTS/InsightGenius/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
celery==5.3.4
redis==5.0.1
httpx==0.25.2
websockets==12.0
plotly==5.15.0
kaleido==0.2.1
python-pptx==0.6.23
reportlab==4.0.7
openai==1.3.6
anthropic==0.7.8
transformers==4.35.2
sentence-transformers==2.2.2
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
